[syft/serde] - move all code related to torch tensor serde into a try-exception block in `third_party.py` - move torch back to under data_science marker in setup.cfg
