+++ packages/syft/setup.cfg
@@ -66,8 +66,6 @@ syft =
     rich==13.7.1
     jinja2==3.1.4
     tenacity==8.3.0
-    # backend.dockerfile installs torch separately, so update the version over there as well!
-    torch==2.3.0
 
 install_requires =
     %(syft)s
@@ -88,6 +86,8 @@ data_science =
     opendp==0.9.2
     evaluate==0.4.1
     recordlinkage==0.16
+    # backend.dockerfile installs torch separately, so update the version over there as well!
+    torch==2.3.0
 
 dev =
     %(test_plugins)s
